refactor(coding-agent): extract gh pr module
Move the PR/issue/repo view + PR write-op domain (getOrFetch* cache-aware fetchers, unified-diff parser, checkout/push/create, view formatters) into gh-pr; gh.ts is now a thin facade: GithubTool dispatch plus full re-export of the public surface.

Op: compress
